@@ -1,26 +1,18 @@
 {
   "name": "beefweb-client",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "description": "Beefweb client library for browser and node",
-  "main": "dist/cjs/index.js",
-  "module": "src/index.js",
+  "main": "src/index.js",
+  "type": "module",
   "repository": "https://github.com/hyperblast/beefweb",
   "author": "Hyperblast",
   "license": "MIT",
   "private": true,
   "scripts": {
-    "build": "rimraf dist && webpack-cli && babel src -d dist/cjs",
-    "build:cjs": "rimraf dist/cjs && babel src -d dist/cjs",
-    "build:umd": "rimraf dist/umd && webpack-cli"
+    "build": "webpack-cli"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.25.0",
-    "babel-loader": "^7.1.1",
-    "babel-preset-env": "^1.7.0",
-    "rimraf": "^2.6.2",
-    "webpack": "^4.16.5",
-    "webpack-bundle-analyzer": "^3.3.2",
-    "webpack-cli": "^3.1.0"
+    "webpack": "^5.75.0",
+    "webpack-cli": "^4.10.0"
   }
 }
