@@ -3,6 +3,7 @@
   "version": "0.7.0",
   "description": "Beefweb user interface",
   "main": "src/index.js",
+  "type": "module",
   "private": true,
   "repository": "https://github.com/hyperblast/beefweb",
   "author": "Hyperblast",
@@ -12,33 +13,35 @@
     "watch": "webpack-cli --watch"
   },
   "dependencies": {
-    "detect-it": "^3.0.4",
+    "detect-it": "^4.0.1",
     "lodash": "^4.17.21",
-    "navigo": "^7.0.0",
-    "normalize.css": "^8.0.0",
+    "navigo": "^7.1.3",
+    "normalize.css": "^8.0.1",
     "open-iconic": "^1.1.1",
-    "prop-types": "^15.5.10",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "react-modal": "^3.1.10",
-    "react-sortable-hoc": "^0.8.3",
-    "wolfy87-eventemitter": "^5.2.2"
+    "prop-types": "^15.8.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-modal": "^3.16.1",
+    "react-sortable-hoc": "^2.0.0",
+    "wolfy87-eventemitter": "^5.2.9"
   },
   "devDependencies": {
-    "babel-core": "^6.25.0",
-    "babel-loader": "^7.1.1",
-    "babel-preset-react": "^6.24.1",
-    "css-loader": "^1.0.0",
-    "extract-text-webpack-plugin": "^4.0.0-beta.0",
-    "file-loader": "^1.1.6",
-    "html-webpack-plugin": "^3.2.0",
-    "less": "^3.8.1",
-    "less-loader": "^4.0.5",
-    "qunit": "^2.4.1",
-    "style-loader": "^0.22.1",
-    "url-loader": "^1.1.1",
-    "webpack": "^4.16.5",
-    "webpack-bundle-analyzer": "^3.3.2",
-    "webpack-cli": "^3.1.0"
+    "@babel/core": "^7.20.12",
+    "@babel/preset-react": "^7.18.6",
+    "babel-loader": "^8.3.0",
+    "css-loader": "^6.7.3",
+    "css-minimizer-webpack-plugin": "^3.4.1",
+    "mini-css-extract-plugin": "^2.7.2",
+    "file-loader": "^6.2.0",
+    "html-webpack-plugin": "^5.5.0",
+    "less": "^4.1.3",
+    "less-loader": "^10.2.0",
+    "terser-webpack-plugin": "5.3.6",
+    "qunit": "^2.19.3",
+    "style-loader": "^3.3.1",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.75.0",
+    "webpack-bundle-analyzer": "^4.7.0",
+    "webpack-cli": "^4.10.0"
   }
 }
