@@ -1,3 +1,3 @@
-export { ErrorType, PlaybackState, SwitchParam } from './enums'
-export { formatQueryString } from './utils'
-export { default as PlayerClient } from './player_client'
+export { ErrorType, PlaybackState, SwitchParam } from './enums.js'
+export { formatQueryString } from './utils.js'
+export { default as PlayerClient } from './player_client.js'
