@@ -1,12 +1,12 @@
-import DataSource from './data_source'
-import PlayerModel from './player_model'
-import PlaylistModel from './playlist_model'
-import FileBrowserModel from './file_browser_model'
-import SettingsModel from './settings_model'
-import NotificationModel from './notification_model';
-import ScrollManager from './scroll_manager';
-import NavigationModel from './navigation_model';
-import ColumnsSettingsModel from './columns_settings_model';
+import DataSource from './data_source.js'
+import PlayerModel from './player_model.js'
+import PlaylistModel from './playlist_model.js'
+import FileBrowserModel from './file_browser_model.js'
+import SettingsModel from './settings_model.js'
+import NotificationModel from './notification_model.js'
+import ScrollManager from './scroll_manager.js'
+import NavigationModel from './navigation_model.js';
+import ColumnsSettingsModel from './columns_settings_model.js';
 
 export default class AppModel
 {
