@@ -1,12 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types'
-import NavigationModel, { SettingsView } from './navigation_model';
-import ModelBinding from './model_binding';
-import GeneralSettings from './general_settings';
-import SettingsModel from './settings_model';
-import ColumnsSettings from './columns_settings';
-import ColumnsSettingsModel from './columns_settings_model';
-import ServiceContext from './service_context';
+import NavigationModel, { SettingsView } from './navigation_model.js';
+import ModelBinding from './model_binding.js';
+import GeneralSettings from './general_settings.js';
+import SettingsModel from './settings_model.js';
+import ColumnsSettings from './columns_settings.js';
+import ColumnsSettingsModel from './columns_settings_model.js';
+import ServiceContext from './service_context.js';
 
 class SettingsContent extends React.PureComponent
 {
